+++ src/components/CarouselButton/CarouselButton.js
@@ -6,11 +6,55 @@ import PreviousChevron from "../SVG/PreviousChevron";
 
 const CarouselButtonElement = styled.button`
   position: absolute;
-  top: 0;
   ${props => (props.previous ? "left: 10px;" : "right: 10px;")};
   z-index: 9;
-  top: 50%;
-  transform: translateY(-50%) scale(1);
+  top: ${({ options }) => {
+    switch (options.position) {
+      case "top":
+        return "10px";
+      case "bottom":
+        return "calc(100% - (42px + 1rem))";
+      default:
+        return "50%";
+    }
+  }};
+
+  ${({ options }) => {
+    switch (options.shape) {
+      case "circle": {
+        if (options.width === "auto") {
+          console.log("You haven't specified a width for the button!");
+        }
+
+        return `width: ${options.width};
+        height: ${options.width};
+        border-radius: 50%;`;
+      }
+      case "square": {
+        if (options.width === "auto") {
+          console.log("You haven't specified a width for the button!");
+        }
+
+        return `width: ${options.width};
+        height: ${options.width};
+        border-radius: 0;`;
+      }
+      default:
+        return `
+            width: ${options.width};
+            height: ${options.height};
+            border-radius: ${options.borderRadius};
+          `;
+    }
+  }}
+
+  ${({ options }) => {
+    if (options.position === "middle") {
+      return "transform: translateY(-50%) scale(1);";
+    }
+  }}
+
+
   background: rgba(0, 0, 0, 0.75);
   border: none;
   padding: 0;
@@ -19,19 +63,25 @@ const CarouselButtonElement = styled.button`
   justify-content: center;
   align-items: center;
   padding: 0.5rem 0;
-  border-radius: 3px;
   transition: all 0.3s ease;
   box-shadow: 0 0 3px rgba(0, 0, 0, 0.16), 0 0 3px rgba(0, 0, 0, 0.23);
 
   &:hover {
     cursor: pointer;
-    transform: translateY(-50%) scale(1.05);
+
+    ${({ options }) => {
+      if (options.position === "middle") {
+        return "transform: scale(1.05);";
+      } else {
+        return "transform: scale(1.05)";
+      }
+    }}
   }
 `;
 
 const CarouselButton = ({ previous }) => {
   const { state, dispatch } = useContext(store);
-  const { activeItem, childCount } = state;
+  const { activeItem, childCount, controlsOptions } = state;
 
   const handleNext = () => {
     dispatch({
@@ -58,14 +108,44 @@ const CarouselButton = ({ previous }) => {
 
   const renderButtonBody = () => {
     if (previous) {
-      return <PreviousChevron width={32} height={32} fill="#ffffff" />;
+      if (
+        controlsOptions.leftControlIcon &&
+        React.isValidElement(controlsOptions.leftControlIcon)
+      ) {
+        return controlsOptions.leftControlIcon;
+      }
+
+      return (
+        <PreviousChevron
+          width={controlsOptions.iconWidth}
+          height={controlsOptions.iconWidth}
+          fill={controlsOptions.iconFill}
+        />
+      );
     }
 
-    return <NextChevron width={32} height={32} fill="#ffffff" />;
+    if (
+      controlsOptions.rightControlIcon &&
+      React.isValidElement(controlsOptions.rightControlIcon)
+    ) {
+      return controlsOptions.leftControlIcon;
+    }
+
+    return (
+      <NextChevron
+        width={controlsOptions.iconWidth}
+        height={controlsOptions.iconWidth}
+        fill={controlsOptions.iconFill}
+      />
+    );
   };
 
   return (
-    <CarouselButtonElement onClick={handleClick} previous={previous}>
+    <CarouselButtonElement
+      onClick={handleClick}
+      previous={previous}
+      options={controlsOptions}
+    >
       {renderButtonBody()}
     </CarouselButtonElement>
   );